Refactor: components -> src 패키지로 이동
