+++ src/app/layout.tsx
@@ -1,8 +1,9 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import Header from "@/components/layout/Header";
-import Footer from "@/components/layout/Footer";
+import Header from "@/src/components/layout/Header";
+import Footer from "@/src/components/layout/Footer";
+import {CartProvider} from "@/src/components/features/home/context/CartContext";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -18,8 +19,6 @@ export const metadata: Metadata = {
   title: "Create Next App",
   description: "Generated by create next app",
 };
-
-import { CartProvider } from "@/components/features/home/context/CartContext";
 
 export default function RootLayout({
   children,
