+++ src/app/member/signup/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { useRouter } from "next/navigation";
-import { SignupModal } from "@/components/features/modals/SignupModal";
+import {SignupModal} from "@/src/components/features/modals/SignupModal";
 
 export default function SignupPage() {
   const router = useRouter();
