+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import ProductSection from "@/components/features/home/ProductSection";
+import ProductSection from "@/src/components/features/home/ProductSection";
 
 export default function Home() {
   return (
