+++ src/components/features/mypage/AddressPanel.tsx
@@ -1,9 +1,9 @@
 "use client";
 
 import { useState } from "react";
-import Button from "@/components/common/Button";
+import Button from "@/src/components/common/Button";
 import { Plus, Pencil, Trash2, Star } from "lucide-react";
-import ConfirmModal from "@/components/features/modals/ConfirmModal";
+import ConfirmModal from "@/src/components/features/modals/ConfirmModal";
 
 interface Address {
   id: number;
